third_party/blink/web_tests/flag-specific/highdpi/printing/page-break-inside-avoid-expected.txt

Test for page-break-inside:avoid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL: expected page number of "page1-1" is 1. Was 0
PASS: page number of "page1-2" is 1
PASS: page number of "page2" is 2
FAIL: expected page number of "page3" is 3. Was 2
FAIL: expected page number of "page4" is 4. Was 3
FAIL: expected page number of "page5" is 5. Was 4
FAIL: expected page number of "page8" is 8. Was 7
FAIL: expected page number of "page9-1" is 9. Was 7
FAIL: expected page number of "page9-2" is 9. Was 8
FAIL: expected page number of "page10" is 10. Was 8

PASS successfullyParsed is true

TEST COMPLETE
